third_party/WebKit/LayoutTests/fast/css/background-repeat-null-y-crash.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<body>
<script>
  document.body.style.backgroundRepeat='repeat';
  document.body.style.backgroundRepeatY='';
  shouldBe(document.body.style.background, '');
</script>
</body>
